+++ java/shared/java/org/signal/libsignal/protocol/message/SignalMessage.java
@@ -7,11 +7,9 @@ package org.signal.libsignal.protocol.message;
 
 import static org.signal.libsignal.internal.FilterExceptions.filterExceptions;
 
-import javax.crypto.spec.SecretKeySpec;
 import org.signal.libsignal.internal.CalledFromNative;
 import org.signal.libsignal.internal.Native;
 import org.signal.libsignal.internal.NativeHandleGuard;
-import org.signal.libsignal.protocol.IdentityKey;
 import org.signal.libsignal.protocol.InvalidKeyException;
 import org.signal.libsignal.protocol.InvalidMessageException;
 import org.signal.libsignal.protocol.InvalidVersionException;
@@ -64,28 +62,6 @@ public class SignalMessage extends NativeHandleGuard.SimpleOwner
 
   public byte[] getPqRatchet() {
     return filterExceptions(() -> guardedMapChecked(Native::SignalMessage_GetPqRatchet));
-  }
-
-  public void verifyMac(
-      IdentityKey senderIdentityKey, IdentityKey receiverIdentityKey, SecretKeySpec macKey)
-      throws InvalidMessageException, InvalidKeyException {
-    try (NativeHandleGuard guard = new NativeHandleGuard(this);
-        NativeHandleGuard senderIdentityGuard =
-            new NativeHandleGuard(senderIdentityKey.getPublicKey());
-        NativeHandleGuard receiverIdentityGuard =
-            new NativeHandleGuard(receiverIdentityKey.getPublicKey()); ) {
-      if (!filterExceptions(
-          InvalidMessageException.class,
-          InvalidKeyException.class,
-          () ->
-              Native.SignalMessage_VerifyMac(
-                  guard.nativeHandle(),
-                  senderIdentityGuard.nativeHandle(),
-                  receiverIdentityGuard.nativeHandle(),
-                  macKey.getEncoded()))) {
-        throw new InvalidMessageException("Bad Mac!");
-      }
-    }
   }
 
   @Override