+++ RELEASE_NOTES.md
@@ -2,3 +2,4 @@ v0.94.0
 
 - keytrans: Detect version changes sooner
 - Expose PQ session archiving ratio API
+- Remove unused SignalMessage.verifyMac function